rename and split files of typeutil
