+++ knowde/_feature/_shared/typeutil/__init__.py
@@ -1,1 +1,2 @@
-from .typeutils import extract_type, is_nested, is_option, is_optional  # noqa: F401
+from .check import is_nested, is_option, is_optional  # noqa: F401
+from .operate import extract_type  # noqa: F401
