+++ knowde/_feature/_shared/typeutil/check.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from types import NoneType, UnionType
+from types import NoneType
 from typing import Any, TypeGuard, get_args
 
 from pydantic import BaseModel
@@ -24,17 +24,6 @@ def is_option(annotation: type[Any] | None) -> bool:
     if is_nested(annotation):
         return False
     return is_optional(annotation)
-
-
-def extract_type(t: type | None | UnionType) -> type:
-    """NoneTypeを取り除いて返す."""
-    if t is None:
-        msg = f"{t} must be type"
-        raise ValueError(msg)
-    args = get_args(t)
-    if NoneType in args:
-        return next(filter(lambda x: x != NoneType, args))
-    return t
 
 
 # def has_fieldargs(t: type[BaseModel], f: Callable) -> bool:
