+++ knowde/_feature/_shared/typeutil/operate.py
@@ -1,0 +1,15 @@
+from __future__ import annotations
+
+from types import NoneType, UnionType
+from typing import get_args
+
+
+def extract_type(t: type | None | UnionType) -> type:
+    """NoneTypeを取り除いて返す."""
+    if t is None:
+        msg = f"{t} must be type"
+        raise ValueError(msg)
+    args = get_args(t)
+    if NoneType in args:
+        return next(filter(lambda x: x != NoneType, args))
+    return t
